Restructured packaging by moving Matchers to its own package.
new package: com.phasmidsoftware.number.matchers
