+++ src/test/scala/com/phasmidsoftware/number/core/ExpressionMatchersSpec.scala
@@ -1,6 +1,7 @@
 package com.phasmidsoftware.number.core
 
 import com.phasmidsoftware.number.core.Number.one
+import com.phasmidsoftware.number.matchers.MatchLogger
 import org.scalatest.BeforeAndAfter
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should
